refactored code into their own modules
